port SHA-1: 308d91de1c094f20885156e71428cfcfc31b129b
* Update plugins, dependencies, and clean up small code inspection issues


git-tfs-id: [https://tfs.codeplex.com/tfs/TFS05/]$/ZXingNet/trunk;C89794

+++ Source/lib/datamatrix/encoder/DefaultPlacement.cs
@@ -65,7 +65,7 @@ namespace ZXing.Datamatrix.Encoder
 
       public void setBit(int col, int row, bool bit)
       {
-         bits[row * numcols + col] = bit ? (byte)1 : (byte)0;
+         bits[row * numcols + col] = (byte)(bit ? 1 : 0);
       }
 
       public bool hasBit(int col, int row)
